refac: rename `parser` to `parsing`
