+++ src/main/kotlin/parsing/treeToAlt.kt
@@ -1,6 +1,6 @@
-package ekko.parser
+package ekko.parsing
 
-import ekko.parser.EkkoParser.AltContext
+import ekko.parsing.EkkoParser.AltContext
 import ekko.tree.Alt
 import java.io.File
