+++ src/main/kotlin/parsing/treeToExp.kt
@@ -1,14 +1,14 @@
-package ekko.parser
+package ekko.parsing
 
-import ekko.parser.EkkoParser.EAbsContext
-import ekko.parser.EkkoParser.EAppContext
-import ekko.parser.EkkoParser.EDecimalContext
-import ekko.parser.EkkoParser.EGroupContext
-import ekko.parser.EkkoParser.EIntContext
-import ekko.parser.EkkoParser.ELetContext
-import ekko.parser.EkkoParser.EStringContext
-import ekko.parser.EkkoParser.EVarContext
-import ekko.parser.EkkoParser.ExpContext
+import ekko.parsing.EkkoParser.EAbsContext
+import ekko.parsing.EkkoParser.EAppContext
+import ekko.parsing.EkkoParser.EDecimalContext
+import ekko.parsing.EkkoParser.EGroupContext
+import ekko.parsing.EkkoParser.EIntContext
+import ekko.parsing.EkkoParser.ELetContext
+import ekko.parsing.EkkoParser.EStringContext
+import ekko.parsing.EkkoParser.EVarContext
+import ekko.parsing.EkkoParser.ExpContext
 import ekko.tree.EAbs
 import ekko.tree.EApp
 import ekko.tree.EGroup
