+++ src/main/kotlin/parsing/treeToPat.kt
@@ -1,7 +1,7 @@
-package ekko.parser
+package ekko.parsing
 
-import ekko.parser.EkkoParser.PVarContext
-import ekko.parser.EkkoParser.PatContext
+import ekko.parsing.EkkoParser.PVarContext
+import ekko.parsing.EkkoParser.PatContext
 import ekko.tree.PVar
 import ekko.tree.Pat
 import java.io.File
